+++ packages/typir/src/kinds/function/function-initializer.ts
@@ -4,17 +4,16 @@
  * terms of the MIT License, which is available in the project root.
  ******************************************************************************/
 
-import { isType, Type, TypeStateListener } from '../../graph/type-node.js';
+import { Type, TypeStateListener } from '../../graph/type-node.js';
 import { TypeInitializer } from '../../initialization/type-initializer.js';
-import { CompositeTypeInferenceRule, InferenceProblem } from '../../services/inference.js';
-import { ValidationRuleStateless } from '../../services/validation.js';
+import { CompositeTypeInferenceRule } from '../../services/inference.js';
 import { TypirServices } from '../../typir.js';
-import { InferenceRuleWithOptions, optionsBoundToType, bindInferCurrentTypeRule } from '../../utils/utils-definitions.js';
-import { checkTypeArrays } from '../../utils/utils-type-comparison.js';
-import { assertType, toArray } from '../../utils/utils.js';
+import { RuleRegistry } from '../../utils/rule-registration.js';
+import { bindInferCurrentTypeRule, InferenceRuleWithOptions, optionsBoundToType } from '../../utils/utils-definitions.js';
+import { assertType } from '../../utils/utils.js';
 import { FunctionCallInferenceRule } from './function-inference-call.js';
-import { CreateFunctionTypeDetails, FunctionKind, OverloadedFunctionDetails } from './function-kind.js';
 import { OverloadedFunctionsTypeInferenceRule } from './function-inference-overloaded.js';
+import { CreateFunctionTypeDetails, FunctionKind, OverloadedFunctionDetails } from './function-kind.js';
 import { FunctionType, isFunctionType } from './function-type.js';
 
 /**
@@ -49,17 +48,20 @@ export class FunctionTypeInitializer<LanguageType = unknown> extends TypeInitial
         } else {
             const overloaded: OverloadedFunctionDetails<LanguageType> = {
                 overloadedFunctions: [],
-                inference: this.createInferenceRuleForOverloads(),
+                details: new RuleRegistry(services as TypirServices),
+                inferenceRule: this.createInferenceRuleForOverloads(),
                 sameOutputType: undefined,
             };
             this.kind.mapNameTypes.set(functionName, overloaded);
+            // the "global" validation for function calls needs to update its registration according to added/removed inference rules for calls of added/removed functions
+            overloaded.details.addListener(kind.validatorArgumentsCalls);
         }
 
         // create the new Function type
         this.initialFunctionType = new FunctionType(kind as FunctionKind, typeDetails);
 
         this.inferenceRules = this.createInferenceRules(typeDetails, this.initialFunctionType);
-        this.registerInferenceRules(functionName, undefined);
+        this.registerRules(functionName, undefined);
 
         this.initialFunctionType.addListener(this, true);
     }
@@ -79,18 +81,18 @@ export class FunctionTypeInitializer<LanguageType = unknown> extends TypeInitial
         const readyFunctionType = this.producedType(functionType);
         if (readyFunctionType !== functionType) {
             functionType.removeListener(this);
-            this.deregisterInferenceRules(functionName, undefined);
+            this.deregisterRules(functionName, undefined);
             this.inferenceRules = this.createInferenceRules(this.typeDetails, readyFunctionType);
-            this.registerInferenceRules(functionName, readyFunctionType);
+            this.registerRules(functionName, readyFunctionType);
         } else {
-            this.deregisterInferenceRules(functionName, undefined);
-            this.registerInferenceRules(functionName, readyFunctionType);
+            this.deregisterRules(functionName, undefined);
+            this.registerRules(functionName, readyFunctionType);
         }
 
         // remember the new function for later in order to enable overloaded functions!
-        // const functionName = typeDetails.functionName;
-        const outputTypeForFunctionCalls = this.kind.getOutputTypeForFunctionCalls(readyFunctionType); // output parameter for function calls
         const overloaded = this.kind.mapNameTypes.get(functionName)!;
+        // Have all overloaded functions the same output type?
+        const outputTypeForFunctionCalls = this.kind.getOutputTypeForFunctionCalls(readyFunctionType); // output parameter for function calls
         if (overloaded.overloadedFunctions.length <= 0) {
             // remember the output type of the first function
             overloaded.sameOutputType = outputTypeForFunctionCalls;
@@ -102,10 +104,15 @@ export class FunctionTypeInitializer<LanguageType = unknown> extends TypeInitial
                 overloaded.sameOutputType = undefined;
             }
         }
+        // remember the function type itself
+        overloaded.overloadedFunctions.push(readyFunctionType);
         // register each inference rule for calls of the function
-        this.typeDetails.inferenceRulesForCalls.forEach(rule => overloaded.overloadedFunctions.push({
+        this.typeDetails.inferenceRulesForCalls.forEach(rule => overloaded.details.addRule({
             functionType: readyFunctionType,
             inferenceRuleForCalls: rule,
+        }, {
+            languageKey: rule.languageKey, // the language keys are directly encoded inside these special inference rules for function calls
+            boundToType: readyFunctionType, // these rules are specific for current function type/signature
         }));
     }
 
@@ -117,26 +124,20 @@ export class FunctionTypeInitializer<LanguageType = unknown> extends TypeInitial
         // nothing specific needs to be done for Functions here, since the base implementation takes already care about all relevant stuff
     }
 
-    protected registerInferenceRules(functionName: string, functionType: FunctionType | undefined): void {
+    protected registerRules(functionName: string, functionType: FunctionType | undefined): void {
         for (const rule of this.inferenceRules.inferenceForCall) {
             const overloaded = this.kind.mapNameTypes.get(functionName)!;
-            overloaded.inference.addInferenceRule(rule.rule, optionsBoundToType(rule.options, functionType));
-        }
-        for (const check of this.inferenceRules.validationForCall) {
-            this.kind.services.validation.Collector.addValidationRule(check, { boundToType: functionType });
+            overloaded.inferenceRule.addInferenceRule(rule.rule, optionsBoundToType(rule.options, functionType));
         }
         for (const rule of this.inferenceRules.inferenceForDeclaration) {
             this.kind.services.Inference.addInferenceRule(rule.rule, optionsBoundToType(rule.options, functionType));
         }
     }
 
-    protected deregisterInferenceRules(functionName: string, functionType: FunctionType | undefined): void {
+    protected deregisterRules(functionName: string, functionType: FunctionType | undefined): void {
         for (const rule of this.inferenceRules.inferenceForCall) {
             const overloaded = this.kind.mapNameTypes.get(functionName);
-            overloaded?.inference.removeInferenceRule(rule.rule, optionsBoundToType(rule.options, functionType));
-        }
-        for (const check of this.inferenceRules.validationForCall) {
-            this.kind.services.validation.Collector.removeValidationRule(check, { boundToType: functionType });
+            overloaded?.inferenceRule.removeInferenceRule(rule.rule, optionsBoundToType(rule.options, functionType));
         }
         for (const rule of this.inferenceRules.inferenceForDeclaration) {
             this.kind.services.Inference.removeInferenceRule(rule.rule, optionsBoundToType(rule.options, functionType));
@@ -146,60 +147,18 @@ export class FunctionTypeInitializer<LanguageType = unknown> extends TypeInitial
     protected createInferenceRules(typeDetails: CreateFunctionTypeDetails<LanguageType>, functionType: FunctionType): FunctionInferenceRules<LanguageType> {
         const result: FunctionInferenceRules<LanguageType> = {
             inferenceForCall: [],
-            validationForCall: [],
             inferenceForDeclaration: [],
         };
-        const mapNameTypes = this.kind.mapNameTypes;
 
         for (const rule of typeDetails.inferenceRulesForCalls) {
             // create inference rule for calls of the new function
             result.inferenceForCall.push({
-                rule: new FunctionCallInferenceRule<LanguageType>(typeDetails, rule, functionType, mapNameTypes),
+                rule: new FunctionCallInferenceRule<LanguageType>(typeDetails, rule, functionType, this.kind.mapNameTypes),
                 options: {
                     languageKey: rule.languageKey,
                     // boundToType: ... this property will be specified outside of this method, when this rule is registered
                 },
             });
-
-            // create validation rule which will be applied when this function is called according to the current inference rule for function calls (this includes the assignability of arguments to input parameters)
-            for (const check of toArray(rule.validation)) {
-                // TODO languageKey ??
-                result.validationForCall.push((languageNode, accept, typir) => {
-                    if ((rule.filter === undefined || rule.filter(languageNode)) && (rule.matching === undefined || rule.matching(languageNode))) {
-                        // check the input arguments, required for overloaded functions
-                        const inputArguments = rule.inputArguments(languageNode);
-                        if (inputArguments && inputArguments.length >= 1) {
-                            // this function type might match, to be sure, resolve the types of the values for the parameters and continue to step 2
-                            const overloadInfos = mapNameTypes.get(typeDetails.functionName);
-                            if (overloadInfos && overloadInfos.overloadedFunctions.length >= 2) {
-                                // for overloaded functions: the types of the parameters need to be inferred in order to determine an exact match
-                                // (Note that the short-cut for type inference for function calls, when all overloads return the same output type, does not work here, since the validation here is specific for this single variant!)
-                                // This is also the reason, why the inference rule for calls is not reused here.)
-                                const childTypes: Array<Type | Array<InferenceProblem<LanguageType>>> = inputArguments.map(child => typir.Inference.inferType(child));
-                                const actualInputTypes = childTypes.filter(t => isType(t));
-                                if (childTypes.length === actualInputTypes.length) {
-                                    const expectedInputTypes = typeDetails.inputParameters.map(p => typir.infrastructure.TypeResolver.resolve(p.type));
-                                    // all operands need to be assignable(! not equal) to the required types
-                                    const comparisonConflicts = checkTypeArrays(actualInputTypes, expectedInputTypes,
-                                        (t1, t2) => typir.Assignability.getAssignabilityProblem(t1, t2), true);
-                                    if (comparisonConflicts.length <= 0) {
-                                        // all arguments are assignable to the expected types of the parameters => this function is really called here => validate this call now
-                                        check(languageNode, functionType, accept, typir);
-                                    }
-                                } else {
-                                    // at least one argument could not be inferred
-                                }
-                            } else {
-                                // the current function is not overloaded, therefore, the types of their parameters are not required => save time
-                                check(languageNode, functionType, accept, typir);
-                            }
-                        } else {
-                            // there are no operands to check
-                            check(languageNode, functionType, accept, typir);
-                        }
-                    }
-                });
-            }
         }
 
         // create inference rule for the declaration of the new function
@@ -215,6 +174,5 @@ export class FunctionTypeInitializer<LanguageType = unknown> extends TypeInitial
 
 interface FunctionInferenceRules<LanguageType = unknown> {
     inferenceForCall: Array<InferenceRuleWithOptions<LanguageType>>;
-    validationForCall: Array<ValidationRuleStateless<LanguageType>>;
     inferenceForDeclaration: Array<InferenceRuleWithOptions<LanguageType>>;
 }